+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.0] - 2024-08-07
+
 ### Removed
 
 - Stopped explicitly installing setuptools and wheel. They will be automatically installed by pip into an isolated build environment if they are required for building a package. ([#243](https://github.com/heroku/buildpacks-python/pull/243))
@@ -173,7 +175,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation. ([#3](https://github.com/heroku/buildpacks-python/pull/3))
 
-[unreleased]: https://github.com/heroku/buildpacks-python/compare/v0.13.0...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-python/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/heroku/buildpacks-python/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/heroku/buildpacks-python/compare/v0.12.1...v0.13.0
 [0.12.1]: https://github.com/heroku/buildpacks-python/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/heroku/buildpacks-python/compare/v0.11.0...v0.12.0